feat(lazy): optimize plugin loading for noice and vimtex
Changed the lazy loading event for folke/noice.nvim from "VeryLazy" to
"LazyFile" to adjust when the plugin is loaded. Enabled filetype-based
lazy loading for lervag/vimtex by setting ft to 'tex' and specifying the
'VimtexInverseSearch' command, ensuring the plugin loads only for TeX
files or when the command is invoked.