feat(cmp): add blink-cmp-spell for spell suggestions
Added 'ribru17/blink-cmp-spell' as a dependency and configured it as
a new completion source for nvim-cmp. The spell source is enabled for
markdown and other filetypes, but only activates in Treesitter @spell
captures and is disabled in @nospell captures. This enhances spelling
suggestions contextually during completion, especially in markdown
files.